+++ backend/src/test/java/com/cafein/backend/api/MemberControllerTest.java
@@ -3,6 +3,7 @@ package com.cafein.backend.api;
 import static com.cafein.backend.support.fixture.MemberFixture.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -11,6 +12,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 import com.cafein.backend.api.member.controller.MemberController;
 import com.cafein.backend.api.member.dto.MemberInfoResponseDTO;
@@ -18,8 +20,7 @@ import com.cafein.backend.api.member.dto.MyPageDTO;
 import com.cafein.backend.api.member.service.MemberInfoService;
 import com.cafein.backend.api.member.service.MyPageService;
 
-@ExtendWith(MockitoExtension.class)
-class MemberControllerTest {
+class MemberControllerTest extends ControllerTestSupporter {
 
 	@Mock
 	private MemberInfoService memberInfoService;
@@ -27,30 +28,31 @@ class MemberControllerTest {
 	@Mock
 	private MyPageService myPageService;
 
-	@InjectMocks
-	private MemberController memberController;
-
 	@Test
-	void 회원_정보를_가져온다() {
+	void 회원_정보를_가져온다() throws Exception {
 		given(memberInfoService.getMemberInfo(any()))
 			.willReturn(MEMBER_INFO_RESPONSE_DTO);
 
-		final ResponseEntity<MemberInfoResponseDTO> result = memberController.getMemberInfo(MEMBER_INFO_DTO);
-
-		assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-		assertThat(result.getBody()).isEqualTo(MEMBER_INFO_RESPONSE_DTO);
+		mockMvc(new MemberController(memberInfoService, myPageService))
+			.perform(MockMvcRequestBuilders
+				.get("/api/member/info"))
+			.andExpect(status().isOk())
+			.andExpect(jsonPath("$.memberName").value("홍길동"))
+			.andExpect(jsonPath("$.email").value("test@test.com"));
 	}
 
 	@Test
-	void 마이페이지에_사용하는_회원_정보를_반환한다() {
+	void 마이페이지에_사용하는_회원_정보를_반환한다() throws Exception {
 		MyPageDTO response = MyPageDTO.builder()
 			.reviewCount(5L)
 			.build();
-		given(myPageService.getMyPageDTO(1L)).willReturn(response);
 
-		final ResponseEntity<MyPageDTO> result = memberController.getMyPage(MEMBER_INFO_DTO);
+		given(myPageService.getMyPageDTO(any())).willReturn(response);
 
-		assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-		assertThat(result.getBody()).isEqualTo(response);
+		mockMvc(new MemberController(memberInfoService, myPageService))
+			.perform(MockMvcRequestBuilders
+				.get("/api/member/mypage"))
+			.andExpect(status().isOk())
+			.andExpect(jsonPath("$.reviewCount").value(5L));
 	}
 }
