+++ backend/src/test/java/com/cafein/backend/api/OAuthLoginControllerTest.java
@@ -1,7 +1,6 @@
 package com.cafein.backend.api;
 
 import static com.cafein.backend.support.fixture.LoginFixture.*;
-import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
@@ -9,7 +8,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MvcResult;
 
 import com.cafein.backend.domain.member.constant.MemberType;
